Refactor: Introduce core structure with app entry, models, and storage services
- Moved `app_entry.dart` to `core` directory and updated import paths.
- Added `AppEntry` widget to manage authentication state and routing.
- Created models for user data and color preferences with JSON serialization.
- Implemented `StorageService` interface and its `SharedPrefsStorage` and `SecureStorageService` implementations for user data management.
- Added validation utilities for email and password.
- Developed color customization logic and UI components for user interaction.
- Established navigation routes using AutoRouter for better structure and maintainability.
