+++ lib/core/models/models.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+  
+
 
 part 'models.freezed.dart';
 part 'models.g.dart';
