+++ lib/core/services/hashing_service.dart
