+++ lib/core/widgets/splash_screen.dart
