+++ lib/features/auth/widgets/logo_image.dart
