+++ lib/features/auth/widgets/password_info.dart
