+++ lib/features/colors/widgets/custom_app_bar.dart
@@ -1,7 +1,7 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:color_aap/core/navigation/navigation.gr.dart';
+import 'package:color_aap/core/services/shared_prefs_storage.dart';
 import 'package:color_aap/generated/l10n.dart';
-import 'package:color_aap/src/services/navigation.gr.dart';
-import 'package:color_aap/src/services/shared_prefs_storage.dart';
 import 'package:flutter/material.dart';
 
 /// Custom app bar widget with user management actions
